backend finished

+++ src/main/java/rs/raf/demo/controllers/UserRestController.java
@@ -2,7 +2,9 @@ package rs.raf.demo.controllers;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+import rs.raf.demo.model.Permissions;
 import rs.raf.demo.model.User;
 import rs.raf.demo.services.UserService;
 
@@ -22,12 +24,31 @@ public class UserRestController {
 
     @GetMapping(value = "/all",
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public List<User> getAllStudents(){
-        return userService.findAll();
+    public ResponseEntity <?> getAllUsers(){
+        if (!authCheck()) {
+            return ResponseEntity.internalServerError().body(null);
+        }
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(Permissions.can_read_users)) {
+            return ResponseEntity.badRequest().build();
+            //return ResponseEntity.status(403).build();
+        }
+
+        List<User> users = userService.findAll();
+        return ResponseEntity.ok(users);
     };
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<?> getUserById(@RequestParam("studentId") Long id){
+    public ResponseEntity<?> getUserById(@RequestParam("id") Long id){
+        if (!authCheck()) {
+            return ResponseEntity.internalServerError().body(null);
+        }
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(Permissions.can_read_users)) {
+            return ResponseEntity.badRequest().build();
+            //return ResponseEntity.status(403).build();
+        }
+
         Optional<User> optionalUser = userService.findById(id);
         if(optionalUser.isPresent()) {
             return ResponseEntity.ok(optionalUser.get());
@@ -38,26 +59,47 @@ public class UserRestController {
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public User createUser(@RequestBody User user){
-        return userService.save(user);
+    public ResponseEntity<?> createUser(@RequestBody User user){
+        if (!authCheck()) {
+            return ResponseEntity.internalServerError().body(null);
+        }
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(Permissions.can_create_users)) {
+            return ResponseEntity.badRequest().build();
+            //return ResponseEntity.status(403).build();
+        }
+
+        return ResponseEntity.ok(userService.save(user));
     }
 
     @PutMapping(consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public User updateUser(@RequestBody User user){
+    public ResponseEntity<?> updateUser(@RequestBody User user){
         System.out.println("updating");
-        return userService.save(user);
+        if (!authCheck()) {
+            return ResponseEntity.internalServerError().body(null);
+        }
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(Permissions.can_update_users)) {
+            return ResponseEntity.badRequest().build();
+            //return ResponseEntity.status(403).build();
+        }
+
+        return ResponseEntity.ok(userService.save(user));
     }
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+        if (!authCheck()) {
+            return ResponseEntity.internalServerError().body(null);
+        }
+        if (!SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(Permissions.can_delete_users)) {
+            return ResponseEntity.badRequest().build();
+        }
         Optional<User> optionalUser = userService.findById(id);
         if(optionalUser.isPresent()) {
             User user = optionalUser.get();
-
-//            for (int i = 0; i < user.getPermissions().size(); i++) {
-//                user.getPermissions().get(i).removeUser(user);
-//            }
+            for (int i = 0; i < user.getPermissions().size(); i++) {
+                user.getPermissions().get(i).removeUser(user);
+            }
             userService.deleteById(id);
 
             return ResponseEntity.ok().build();
@@ -65,5 +107,12 @@ public class UserRestController {
         return ResponseEntity.notFound().build();
     }
 
+    private boolean authCheck(){
+        if (SecurityContextHolder.getContext() == null ||
+                SecurityContextHolder.getContext().getAuthentication() == null ||
+                SecurityContextHolder.getContext().getAuthentication().getAuthorities() == null)
+            return false;
+        return true;
+    }
 
 }
